requirements.txt

streamlit
langgraph
langchain-google-genai
python-dotenv
pandas
requests
pymupdf
scikit-learn
transformers
deepface
opencv-python
speechrecognition
reportlab
tf-keras 
# tf-keras is often needed for DeepFace/Transformers compatibility
